part 10 done...

+++ domain/src/main/java/mg/amas/domain/usecase/CartSummaryUseCase.kt
@@ -5,5 +5,5 @@ import mg.amas.domain.repository.CartRepository
 class CartSummaryUseCase(
     private val repository: CartRepository,
 ) {
-    suspend fun execute(userId: Int) = repository.getCartSummary(userId)
+    suspend fun execute(userId: Long) = repository.getCartSummary(userId)
 }
